US110605 Followup: Serge run_command
Based on the Serge docs, there's a `run_command` plugin with which you can run arbitrary commands. This adds some commands to build our language files for us, so that (in theory) Serge will return to us both the translated JSON files, as well as built JavaScript files, which we can then import. Importing these by name allows the BSI analyzer to actually include the files correctly - we load only the files we need initially, and lazy-load the others if they're needed (e.g. if the page language changes).

I say "in theory" above as I've not been able to actually test this on the Serge service itself - the commands work as expected locally, but I'm still not 100% sure about how the Serge integration actually works - still tracking down some of those details. For one, I don't know if the last `run_command` will actually generate and commit the files from Serge back to the repo, or...? But yeah, still looking.