Call `associate-new-grade`
This processes and passes the correct info to the activity-usage save to `associate-new-grade`. There is some gross stuff like having to store the assignment `name` in the activity score out of MobX object, but it is required because the `activity` needs to pass it in the save to create a new grade with that name.  The `associate-new-grade` action is in the grade-candidates response. The `score-out-of` does not have access to the grade candidates, so the `associateNewGradeAction` variable is being set when the grades dialog is opened and the "Create and link new grade object" is selected. If the dialog isn't opened, such as if we're adding an activity to grades for the first time, it will default to using the `score-out-of` in the `siren-sdk`